+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## [Unreleased]
 
+## [v2.0.0] - 2018-10-05
+
+### Changed
+
+* Unite source file into src/js/ue-scroll.js
+* Update build tools (webpack 4 and Babel 7)
+* Added Japanese document
+
+### Removed
+
+* Removed src/js/ue-scroll.esm.js (united into src/js/ue-scroll.js)
+
 ## [v1.0.1] - 2018-05-02
 
 ### Fixed
@@ -25,7 +37,8 @@
 
 * First released ue-scroll.js
 
-[Unreleased]: https://github.com/azrsix/ue-scroll/compare/v1.0.1...HEAD
+[Unreleased]: https://github.com/azrsix/ue-scroll/compare/v2.0.0...HEAD
+[v2.0.0]: https://github.com/azrsix/ue-scroll/compare/v1.0.1...v2.0.0
 [v1.0.1]: https://github.com/azrsix/ue-scroll/compare/v1.0.0...v1.0.1
 [v1.0.0]: https://github.com/azrsix/ue-scroll/compare/v0.0.1...v1.0.0
 [v0.0.1]: https://github.com/azrsix/ue-scroll/tree/v0.0.1